clean up data_item handling in TimePlotgui
mechanism to match and update number of devices connected, number of data items present, and number of line settings in settings dictionary is kinda confusing and not easy to follow. I implement here some changes to try to make the mechanism more clear. This includes changes in TimePlotgui, PlotItemSettings, and TimePlotDataItem class

Detailed:
* create functiosn to add, remove, and clear line settings to PlotItemSetting
* create function to create data items in TimePlotGui
* create function to add and remove data times from data table to make code more clear and to avoid repetiion of code since multiple functions add and remove data time (e.g. on initialization and when data is loaded).

the chaos is not solved yet. I will include further changes on that in future commits.