move _open_color_dialog() function to LineSettingsQWidgetAction
Detailed:
So far color dialog functions were part of TimePlotDataItem class. This is not desirable because it mixes functions which are supposed to act as interface between PlotSettings (PlotDataItem) and the actual data (i.e. JSON file) with functionality of generating side windows. To fix this I moved functions related to generating the QColorDialog window to the LineSettingsQWidgetAction class in context_menu.py. This makes more sense since this class already facilitates functions related to the different blocks in the line settings submenu. To achieve this i changed the following things:

* moved _open_color_dialog() to LineSettingsQWidgetAction
* introduced getter and setter function in TimePlotDataItem to allow efficient data transfer
* replaced setter and getter functions in PlotDataItemV2 with functions in TimePlotDataItem
* removed redundant functions
* improved doc strings